Add /import command
