+++ glug-bot/Cargo.toml
@@ -9,3 +9,4 @@ log = "0.4"
 pretty_env_logger = "0.5"
 tokio = { version =  "1.8", features = ["rt-multi-thread", "macros"] }
 glug-glug-core = { path = "../glug-core/" }
+glug-glug-importer = { path = "../glug-importer/" }
