+++ glug-core/src/database/drinks.rs
@@ -1,4 +1,4 @@
-use std::collections::{BTreeSet, HashMap};
+use std::collections::BTreeSet;
 
 use sqlx::{query, query_scalar};
