+++ glug-importer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "glug-importer"
+name = "glug-glug-importer"
 version = "0.1.0"
 edition = "2024"
